Move cc `orchestrator_client.{cc,h}` into a separate sdk package
