+++ cc/containers/hello_world_trusted_app/main.cc
@@ -20,13 +20,12 @@
 #include "absl/status/statusor.h"
 #include "app_service.h"
 #include "cc/containers/hello_world_trusted_app/app_service.h"
-#include "cc/containers/hello_world_trusted_app/orchestrator_client.h"
+#include "cc/containers/sdk/orchestrator_client.h"
 #include "grpcpp/security/server_credentials.h"
 #include "grpcpp/server.h"
 #include "grpcpp/server_builder.h"
-#include "orchestrator_client.h"
 
-using ::oak::oak_containers_hello_world_trusted_app::OrchestratorClient;
+using ::oak::containers::sdk::OrchestratorClient;
 using ::oak::oak_containers_hello_world_trusted_app::TrustedApplicationImpl;
 
 int main(int argc, char* argv[]) {
